refactor: remove dead code (VoxelMap)
